add name for call api. Try something for cache logout
