+++ front/src/utils/apolloclient.ts
@@ -1,0 +1,25 @@
+import { ApolloClient, createHttpLink, InMemoryCache } from '@apollo/client';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { setContext } from '@apollo/client/link/context';
+
+const httpLink = createHttpLink({
+	uri: process.env.REACT_APP_API_DEV,
+});
+
+const authLink = setContext((_, { headers }) => {
+	const token = localStorage.getItem('token');
+	return {
+		headers: {
+			...headers,
+			authorization: token,
+		},
+	};
+});
+
+const client = new ApolloClient({
+	uri: process.env.REACT_APP_API_DEV,
+	cache: new InMemoryCache(),
+	link: authLink.concat(httpLink),
+});
+
+export default client;
